level/src/requirements.txt
Django==4.0.3
numpy==1.22.3
djangorestframework==3.14.0
